+++ source/IAccessibleHandler/__init__.py
@@ -3,7 +3,7 @@
 # This file is covered by the GNU General Public License.
 # See the file COPYING for more details.
 
-from typing import Tuple
+import re
 import struct
 import weakref
 from ctypes import (
@@ -21,97 +21,28 @@ from ctypes.wintypes import HANDLE
 from comtypes import IUnknown, IServiceProvider, COMError
 import comtypes.client
 import oleacc
+import JABHandler
 import UIAHandler
 
 from comInterfaces import Accessibility as IA
 
 from comInterfaces import IAccessible2Lib as IA2
-import config
-
-
-_winEventNameCache = {}
-
-
-def getWinEventName(eventID):
-	""" Looks up the name of an EVENT_* winEvent constant. """
-	global _winEventNameCache
-	if not _winEventNameCache:
-		_winEventNameCache = {y: x for x, y in vars(winUser).items() if x.startswith('EVENT_')}
-		_winEventNameCache.update({y: x for x, y in vars(IA2).items() if x.startswith('IA2_EVENT_')})
-	name = _winEventNameCache.get(eventID)
-	if not name:
-		name = "unknown event ({eventID})"
-	return name
-
-
-_objectIDNameCache = {}
-
-
-def getObjectIDName(objectID):
-	""" Looks up the name of an OBJID_* winEvent constant. """
-	global _objectIDNameCache
-	if not _objectIDNameCache:
-		_objectIDNameCache = {y: x for x, y in vars(winUser).items() if x.startswith('OBJID_')}
-	name = _objectIDNameCache.get(objectID)
-	if not name:
-		name = str(objectID)
-	return name
-
-
-def getWinEventLogInfo(window, objectID, childID, eventID=None, threadID=None):
-	"""
-	Formats the given winEvent parameters into a printable string.
-	window, objectID and childID are mandatory,
-	but eventID and threadID are optional.
-	"""
-	windowClassName = winUser.getClassName(window) or "unknown"
-	objectIDName = getObjectIDName(objectID)
-	processID = winUser.getWindowThreadProcessID(window)[0]
-	if processID:
-		processName = appModuleHandler.getAppModuleFromProcessID(processID).appName
-	else:
-		processName = "unknown application"
-	messageList = []
-	if eventID is not None:
-		eventName = getWinEventName(eventID)
-		messageList.append(f"{eventName}")
-	messageList.append(
-		f"window {window} ({windowClassName}), objectID {objectIDName}, childID {childID}, "
-		f"process {processID} ({processName})"
-	)
-	if threadID is not None:
-		messageList.append(f"thread {threadID}")
-	return ", ".join(messageList)
-
-
-def isMSAADebugLoggingEnabled():
-	""" Whether the user has configured NVDA to log extra information about MSAA events. """
-	return config.conf["debugLog"]["MSAA"]
-
-
-IAccessibleObjectIdentifierType = Tuple[
-	int,  # windowHandle
-	int,  # objectID
-	int,  # childID
-]
-
-from . import internalWinEventHandler
-
-from logHandler import log
-import JABHandler
-import eventHandler
-import winUser
 import api
+import appModuleHandler
+import controlTypes
+import core
+import eventHandler
+import keyboardHandler
+from logHandler import log
+import mouseHandler
 import NVDAObjects.IAccessible
 import NVDAObjects.window
-import appModuleHandler
-import mouseHandler
-import controlTypes
-import keyboardHandler
-import core
-import re
+import winUser
 
+from . import internalWinEventHandler
 from .orderedWinEventLimiter import MENU_EVENTIDS
+from .utils import getWinEventLogInfo, getWinEventName, isMSAADebugLoggingEnabled
+
 
 # Special Mozilla gecko MSAA constant additions
 NAVRELATION_LABEL_FOR = 0x1002
