+++ source/IAccessibleHandler/internalWinEventHandler.py
@@ -13,8 +13,7 @@ from typing import Dict, Callable
 
 import core
 import winUser
-from . import getWinEventLogInfo
-from . import isMSAADebugLoggingEnabled
+from .utils import getWinEventLogInfo, isMSAADebugLoggingEnabled
 
 from comInterfaces import IAccessible2Lib as IA2
