+++ source/IAccessibleHandler/orderedWinEventLimiter.py
@@ -3,9 +3,9 @@ import heapq
 import itertools
 
 import winUser
-from . import IAccessibleObjectIdentifierType
+from .types import IAccessibleObjectIdentifierType
 from logHandler import log
-from . import isMSAADebugLoggingEnabled, getWinEventLogInfo
+from .utils import getWinEventLogInfo, isMSAADebugLoggingEnabled
 
 
 MAX_WINEVENTS_PER_THREAD = 10
