+++ user_docs/en/changes.t2t
@@ -141,6 +141,7 @@ What's New in NVDA
    - `speech.speechMode_off` becomes `speech.SpeechMode.off`
    - `speech.speechMode_beeps` becomes `speech.SpeechMode.beeps`
    - `speech.speechMode_talk` becomes `speech.SpeechMode.talk`
+- `IAccessibleHandler.IAccessibleObjectIdentifierType` is now `IAccessibleHandler.types.IAccessibleObjectIdentifierType`. (#12367)
 
 
 = 2020.4 =
